Add "package-for-deploy" shortcut command to package.json
Run as:

 npm run-script package-for-deploy

Note that it removes `aws-sdk` and it's dependencies from node_modules/
because this module is already provided in the lamdba environment and
the resulting "zip" file size drops from about 28MB to 100K. (!!)

After running this, you may want to run "npm install" to bring back
the aws-sdk.

This approach is imperfect in that it installs the /latest/ version of
the dependencies, which may differ from the versions specified in
package.json.